Update the react peerdep to not peg to React 17
This package is currently pegged by peer dependency to React 17, which is likely to be the least adopted React version in a while, given that it was a glorified minor version before React 18. This peer dep should be more than fine for what this package does, and just increases compatibility without causing problems for users that are now able to use the package where they wouldn't be allowed to before.