fix: keystone dep breaks yarn 4 installation
If you try to use web3-onboard with yarn v4 (although I think the
problem will be there with yarn v2 and v3) it fails with:
YN0001: │ Error: @keystonehq/eth-keyring@^0.14.00.3 isn't supported by
any available resolver

First install is ok, but if you run the install again you’ll get the
above. 0.14.00.3 is apparently not a valid range and the only way to
get this fixed is to fix the dependency. Yarn v1 resolves
0.14.00.3 to 0.14.4 and that is the version I’ve set here.